+++ src/main/java/study/domain/Calculator.java
@@ -1,4 +1,11 @@
-package study;
+/**
+ * Calculator.java
+ * 계산 작업을 하는 클래스
+ * 우아한테크코스 1주차 미션 온보딩 - 문자열 계산기
+ * 히로-핀
+ * https://github.com/hotheadfactory/java-calculator
+ */
+package study.domain;
 
 import java.util.Arrays;
 import java.util.LinkedList;
